chore: fix ci
