+++ examples/native/src/SlowList.test.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import { render, screen } from '@testing-library/react-native';
 import { expect, test } from '@jest/globals';
 import { SlowList } from './SlowList';
