+++ examples/web-vite/src/SlowList.test.tsx
@@ -1,3 +1,5 @@
+// @ts-expect-error Needed for Jest testing
+import * as React from 'react';
 import { render, screen } from '@testing-library/react';
 import { SlowList } from './SlowList';
