+++ examples/web-vite/src/SlowList.tsx
@@ -21,10 +21,10 @@ interface ItemProps {
 }
 
 const SlowListItem = ({ title }: ItemProps) => {
-  const [, forceRender] = React.useState<{}>();
+  const [, forceRender] = React.useState();
 
   React.useEffect(() => {
-    forceRender({});
+    forceRender(undefined);
   }, [title]);
 
   return (
